+++ src/main/java/nextstep/member/MemberDao.java
@@ -1,6 +1,6 @@
 package nextstep.member;
 
-import nextstep.infrastructure.role.Role;
+import nextstep.infrastructure.template.Role;
 import nextstep.support.exception.auth.NoSuchMemberException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
